DHSCFT-145: Add benchmark colours to the bar chart when comparing areas
