+++ frontend/fingertips-frontend/components/views/OneIndicatorTwoOrMoreAreasView/index.tsx
@@ -57,6 +57,7 @@ export default async function OneIndicatorTwoOrMoreAreasView({
               {
                 indicatorId: Number(indicatorSelected[0]),
                 areaCodes: [...requestAreas],
+                comparisonMethod: 'Rag',
               },
               API_CACHE_CONFIG
             )
