+++ frontend/fingertips-frontend/lib/chartHelpers/chartHelpers.ts
@@ -3,6 +3,11 @@ import {
   HealthDataPoint,
 } from '@/generated-sources/ft-api-client';
 import { areaCodeForEngland } from './constants';
+import { getBenchmarkTagStyle } from '@/components/organisms/BenchmarkLabel/BenchmarkLabelConfig';
+import {
+  BenchmarkLabelGroupType,
+  BenchmarkLabelType,
+} from '@/components/organisms/BenchmarkLabel/BenchmarkLabelTypes';
 
 export function sortHealthDataForAreasByDate(
   data: HealthDataForArea[]
@@ -77,3 +82,12 @@ export function getMostRecentData(
 export async function loadHighchartsModules(callback: () => void) {
   await import('highcharts/highcharts-more').then(callback);
 }
+
+export const getBenchmarkColour = (benchmarkComparison: BenchmarkLabelType) => {
+  const colours = getBenchmarkTagStyle(
+    BenchmarkLabelGroupType.RAG,
+    benchmarkComparison
+  );
+  const backgroundColor = colours?.backgroundColor;
+  return backgroundColor === 'transparent' ? undefined : backgroundColor;
+};
